Added application to profile
