+++ client/src/components/ApplicationCard/style.module.scss
@@ -1,0 +1,43 @@
+@use '@/styles/vars.scss' as vars;
+
+.responseList {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.responseContainer {
+  max-height: 500px;
+  overflow: hidden;
+  transition: max-height 0.3s ease-in-out;
+
+  &.expanded {
+    max-height: none;
+  }
+}
+
+.question {
+  font-weight: 700;
+  a {
+    color: vars.$blue;
+  }
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.response {
+  padding: 0.75rem;
+
+  &:not(:last-child) {
+    margin-bottom: 1.5rem;
+  }
+}
+
+.submissionHeader {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 2rem;
+}
