+++ client/src/components/Profile/style.module.scss
@@ -8,50 +8,19 @@
   padding: 6rem var(--side-padding);
 }
 
+.profileContainer {
+  display: grid;
+  grid-template-areas:
+    'profile profile'
+    'liability photo'
+    'application application';
+  gap: 2rem;
+}
+
 .profile {
-  margin: 0 auto;
-  max-width: 646px;
-  width: 100%;
+  grid-area: profile;
+}
 
-  .profileHeader {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-
-    .editBtn,
-    .closeBtn {
-      flex: none;
-      cursor: pointer;
-    }
-
-    .closeBtn {
-      margin-left: auto;
-    }
-
-    .logoutBtn {
-      margin-left: auto;
-    }
-  }
-
-  .profileContent {
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-  }
-
-  .buttonGroup {
-    display: flex;
-    justify-content: flex-end;
-    gap: 0.5rem;
-
-    @media screen and (max-width: vars.$breakpoint-md) {
-      .discardButton {
-        display: none;
-      }
-
-      .saveButton {
-        width: 100%;
-      }
-    }
-  }
+.application {
+  grid-area: application;
 }
