+++ client/src/components/ProfileCard/logout.ts
